reporting with missions

+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/reporting/ReportingEntity.kt
@@ -7,7 +7,7 @@ import java.util.UUID
 
 data class ReportingEntity(
     val id: Int? = null,
-    val reportingId: Int? = null,
+    val reportingId: Long? = null,
     val sourceType: SourceTypeEnum? = null,
     val semaphoreId: Int? = null,
     val controlUnitId: Int? = null,
